% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAssumption.R
\name{addAssumption}
\alias{addAssumption}
\title{addAssumption}
\usage{
addAssumption(title, notes, requirements = NA, logged_by = NA)
}
\arguments{
\item{title, }{string: name your assumption for easy reference.}

\item{notes, }{string: notes describing your assumption}

\item{requirements, }{string: any requirements your assumption entails,
e.g. software access. Defaults to NA.}

\item{logged_by, }{string: who logged the assumption? Useful for teams.
Defaults to NA.}
}
\description{
When creating a project plan, you often have to make a number of assumptions.
Record them so they can be added to your project plan document.
}
\examples{

addAssumption("Avatar state", "Don't die while in it", "Not dying", "Kora")


}
