% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{assumptions}
\usage{
assumptions()
}
\value{
assumptions, a dataframe listing, well, assumptions.
}
\description{
Find out which assumptions you and your team have logged about this project
}
\examples{

assumptions()

}
