% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasicDocumentation.R
\name{createBasicDocumentation}
\alias{createBasicDocumentation}
\title{createBasicDocumentation}
\usage{
createBasicDocumentation()
}
\description{
This function will create a basic documentation for all of the data files
in the /data folder for a project.
}
\details{
Currently only operational for .csv files as a start point: more formats soon!

Note, the file will be read as is: you may need to do some cleaning first.
Suggest creating /data/raw and putting raw files there with cleaning script.
}
