% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClient.R
\name{createClient}
\alias{createClient}
\title{createClient}
\usage{
createClient(name, path)
}
\arguments{
\item{name, }{string: name of client.}

\item{path, }{string: path where you want to put the client directory}
}
\description{
One way to organise your projects is by client. Consultthat works this way.
This function creates a new client.
}
\examples{
createClient("RStars", "~/practice")

}
