% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGantt.R
\name{createGantt}
\alias{createGantt}
\title{createGantt}
\usage{
createGantt(key, description, start, end, done, neededBy)
}
\arguments{
\item{key}{give each task a name}

\item{description}{describe the task}

\item{start}{date task should start}

\item{end}{date task should finish}

\item{done}{percentage completion of task}

\item{neededBy}{date completion required}
}
\description{
Create a Gantt chart, using the plan package. You can manipulate a .csv as required.
}
\examples{

task1 <- c("Team Avatar!", "Welcome to Republic City",
as.Date("2018/01/01"), as.Date("2018/01/02"), done = 72, neededBy = as.Date("2018/02/01"))

task2 <- c("The Equalists = bad", "A Leaf in the Wind",
as.Date("2018/01/18"), as.Date("2018/01/24"), done = 65, neededBy = asDate("2018/02/02"))

task3 <- c("Bolin gets captured!", "The Revelation",
as.Date("2018/01/26"), as.Date("2018/01/30"), done = 50, neededBy = asDate("2018/02/05"))

tasks <- rbind(task1, task2, task3)

createGantt(tasks[,1], tasks[,2], tasks[,3], tasks[,4], tasks[,5], tasks[,6])


}
