% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{createProject}
\usage{
createProject(consult_path, client, project_name)
}
\arguments{
\item{consult_path, }{string: path that consulting projects lie on}

\item{client, }{string: name of client}

\item{project_name, }{string: name of project}
}
\description{
Create a package structure for a new consulting project. This builds on the
normal package structure and provides subdirectories for data, outputs (client
facing and internal), meetings etc.
}
\examples{
createProject("~/practice", "RStars", "firstProject")




}
