% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProjectDocuments.R
\name{findDocumentDirectory}
\alias{findDocumentDirectory}
\title{findDocumentDirectory}
\usage{
findDocumentDirectory(project = ".")
}
\arguments{
\item{project}{string: the project directory, default is "."}
}
\value{
the sub-directory as a string
}
\description{
Given a \code{consultthat} project path, attempts to find the location of the
project documents sub-directory.
}
