% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people}
\alias{people}
\title{people()}
\usage{
people(path, client)
}
\arguments{
\item{path, }{path where client directory is located.}

\item{client, }{client you wish to return personnel log for.}
}
\value{
people, a data frame.
}
\description{
Return the log of personnel for the current client as a data frame. If no
log of personnel exists, NULL will be returned and also congratulations you're either
really early in this project or have a great memory. Possibly both.
}
\examples{

people()

}
