% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punchOff.R
\name{punchOff}
\alias{punchOff}
\title{punchOff}
\usage{
punchOff(category, notes)
}
\arguments{
\item{category, }{string: category of work. Divide your work up into 'buckets'
so you can track what part of the project you were working on. E.g. 'data analysis',
'modelling', 'write up', 'meeting'. Need to match category assigned at punchOn().}

\item{notes, }{string: any notes you'd like to add}
}
\description{
Record when you finished working on a project, with notes.
To be used in conjunction with punchOn().
Time sheet will be located in /project/project_documents/time_management.
}
