% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punchOn.R
\name{punchOn}
\alias{punchOn}
\title{punchOn}
\usage{
punchOn(name = NULL, category = NA, notes = NA, project = ".")
}
\arguments{
\item{name, }{string: person who is punching on for this project}

\item{category, }{string: category of work. Divide your work up into 'buckets'
so you can track what part of the project you were working on. E.g. 'data analysis',
'modelling', 'write up', 'meeting'.}

\item{notes, }{string: any notes you'd like to add}

\item{project}{string: where is the project you are punching into}
}
\description{
Record the time you spend on a project.
}
\details{
punchOn will record the system time you started in the current working directory.
The time sheet will be located in the client directory and will record all
work for that client's projects. If the time sheet does not exist, it will create it.

If you are \emph{already} punched on, then you will be punched off and punched back on. This makes
it easy to switch \emph{categories} without punching off first.

Time sheet will be located in /project/project_documents/time_management.

Note: this is tidy time sheet management. Each project can generate a timesheet,
but you can append them all to analyse your overall time management.
Output as a .csv, this is readable for those who may not be using R. You can read
it into R and calculate time spent etc. as required.
}
\examples{

punchOn("Steph", "clean data", "it's a mess")

}
