% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheet.R
\name{timesheet}
\alias{timesheet}
\title{timesheet}
\usage{
timesheet(name, project = ".")
}
\arguments{
\item{name, }{person you'd like to return a timesheet for}

\item{project}{which project you want the timesheet for}
}
\value{
timesheet, a data frame.
}
\description{
Return the timesheet for the current project as a data frame.
If no timesheet exists for this project because you have not punchedIn,
NULL will be returned instead.
}
\examples{
timesheet("Gary")

}
