% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIssue .R
\name{addIssue}
\alias{addIssue}
\title{addIssue}
\usage{
addIssue(ID, category, notes)
}
\arguments{
\item{ID, }{string: choose an ID that you will use to close the issue or make changes to it.
Can be numeric if you wish, or text based. E.g. "00001" or "that damn bug".}

\item{category, }{string: you can choose a category to file it under for later reference
and analysis.}

\item{notes, }{string: it's often worthwile to add some notes to help you when you
pick up the project again. This might be a brief run down of what the problem is,
or it could be a helpful way to store and file useful links, tips etc.}
}
\description{
A simple way to keep track of issues you want to follow up. Add an issue,
it will be added to a .csv which is accessible to non-R users. If no issue
tracker for this project exists, one will be created.
}
\examples{

addIssue("stupid bug 3", "breaking code", "I can't work out where that damn missing bracket is. WHY?!?!")


}
