% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeIssue.R
\name{closeIssue}
\alias{closeIssue}
\title{closeIssue}
\usage{
closeIssue(user, ID, notes = NA)
}
\arguments{
\item{user, }{string: the person who closed the issue.}

\item{ID, }{string: the ID of the issue you want to close.}

\item{notes, }{string: any closing notes you want to add to the issue.}
}
\description{
Close an issue you've previously logged.
Add notes on how you closed it for future reference.
}
