% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{createProject}
\usage{
createProject(consult_path = ".", client = NULL, project_name = NULL,
  overwrite = FALSE, project_documents = c("meetings", "documents",
  "initiation", "financials", "time", "planning", "milestones", "outputs",
  "data", "documentation"), documents_directory = "project_documents",
  use_package = TRUE)
}
\arguments{
\item{consult_path}{string: path that consulting projects lie on, OR full path to directory}

\item{client}{string: name of client}

\item{project_name}{string: name of project}

\item{overwrite}{logical: should existing project_documents directories be overwritten?}

\item{project_documents}{string: the consulting project documents you want in this project, see Details}

\item{documents_directory}{string: what sub-directory should the project documents be in? Default is \code{project_documents}}

\item{use_package}{logical: should a package structure be used, or just create the directory? default is \code{TRUE}}
}
\description{
Create a package structure for a new consulting project. This builds on the
normal package structure and provides subdirectories for data, outputs (client
facing and internal), meetings etc.
}
\examples{
createProject("~/practice", "RStars", "firstProject")




}
