% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues}
\alias{issues}
\title{issues()}
\usage{
issues(project = ".")
}
\arguments{
\item{project}{string: the project to return issues for}
}
\value{
issues, a data frame.
}
\description{
Return the log of issues for the current project as a data frame. If no
log of issues exists, NULL will be returned and also congratulations you're either
really early in this project of amazing. Possibly both.
}
\examples{

issues()

}
